ms_recommender_original/requirements.txt
tensorflow-gpu==1.15.4
scikit-learn>=0.19.1
scikit-surprise>=1.0.6
numpy>=1.13.3
pyyaml>=5.4.1,<6
tqdm==4.62.3
retrying==1.3.3
nltk>=3.4


